fix: comprehensive CUDA Graph safety audit — fix all remaining issues
After 10+ deployments hitting different errors each time, did a full
audit of all code paths during CUDA Graph warmup/capture. Fixes:

1. vllm.py: Replace torch.cuda.is_current_stream_capturing() with
   _graph_intended for T>1 prefill guard. The stream capture check is
   unreliable on vLLM's non-default capture streams.

2. vllm.py: Remove all torch.zeros()/torch.zeros_like() calls from
   graph-reachable paths (T>1 shortcut x2, no-data fallback). These
   allocate new tensors during graph capture which is forbidden.
   Now uses output.zero_() (in-place) or falls through to orig_fn.

3. score.py: Replace torch.zeros() in compute_hybrid_attention() with
   query * 0.0 — produces same-shaped zero tensor without allocation.

4. capture.py: Change _arange_buf from int32 to int64 for consistency
   with _pos_tensor, _count_tensor, _capacity_tensor (all int64 now).
   Update peek_full() docstring to match.

+++ turboquant/capture.py
@@ -171,7 +171,7 @@ class RingBuffer:
             [self.capacity], device=self.device, dtype=torch.int64,
         )
         self._arange_buf = torch.arange(
-            self.capacity, device=self.device, dtype=torch.int32,
+            self.capacity, device=self.device, dtype=torch.int64,
         )
         self._graph_mode = True
 
@@ -212,9 +212,9 @@ class RingBuffer:
             (ring_k, ring_v, count_tensor, arange_buf, capacity_tensor)
             - ring_k: (capacity, H, D) — full buffer
             - ring_v: (capacity, H, D) — full buffer
-            - count_tensor: (1,) int32 — number of valid entries
-            - arange_buf: (capacity,) int32 — for masking
-            - capacity_tensor: (1,) int32 — ring buffer capacity
+            - count_tensor: (1,) int64 — number of valid entries
+            - arange_buf: (capacity,) int64 — for masking
+            - capacity_tensor: (1,) int64 — ring buffer capacity
         """
         return self._k, self._v, self._count_tensor, self._arange_buf, self._capacity_tensor
